b/requirements.txt
fastapi
uvicorn
grpcio
grpcio-tools
aiohttp
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation-aiohttp-client